+++ sdks/java/extensions/google-cloud-platform-core/src/main/java/org/apache/beam/sdk/extensions/gcp/util/GcsUtil.java
@@ -1041,7 +1041,8 @@ public class GcsUtil {
   }
 
   public void remove(Collection<String> filenames) throws IOException {
-    // TODO(BEAM-8268): It would be better to add per-file retries and backoff
+    // TODO(https://github.com/apache/beam/issues/19859): It would be better to add per-file retries
+    // and backoff
     // instead of failing everything if a single operation fails.
     executeBatches(makeRemoveBatches(filenames));
   }