+++ sdks/python/apache_beam/options/pipeline_options.py
@@ -302,8 +302,9 @@ class PipelineOptions(HasDisplayData):
       Dictionary of all args and values.
     """
 
-    # TODO(BEAM-1319): PipelineOption sub-classes in the main session might be
-    # repeated. Pick last unique instance of each subclass to avoid conflicts.
+    # TODO(https://github.com/apache/beam/issues/18197): PipelineOption
+    # sub-classes in the main session might be repeated. Pick last unique
+    # instance of each subclass to avoid conflicts.
     subset = {}
     parser = _BeamArgumentParser()
     for cls in PipelineOptions.__subclasses__():
@@ -962,6 +963,14 @@ class WorkerOptions(PipelineOptions):
             'separated by a comma where first value gives a regex to '
             'identify the container image to override and the second value '
             'gives the replacement container image.'))
+    parser.add_argument(
+        '--default_sdk_harness_log_level',
+        default=None,
+        help=(
+            'Controls the default log level of all loggers without a log level '
+            'override. Values can be either a labeled level or a number '
+            '(See https://docs.python.org/3/library/logging.html#levels). '
+            'Default log level is INFO.'))
     parser.add_argument(
         '--use_public_ips',
         default=None,